chore: update typings
